CLOUDSTACK-XXXX: In network tab - Time from clicking to completely loading of network takes 16 sec
Instead of searching in each network, now it is searching in each zone.
For basic zone, it will show the security group directly, because by default the securitygroupsenabled is true.
For advanced zone, it checks the securitygroupsenabled option in each zone.
If any one has value true, then the security group is shown.