CLOUDSTACK-XXXX: listPortForwardingRules not scaling well
For setting the width of each data item for each row of portforwarding rules, it was processing all rules.
Basically it was sending all rules, which is un-necessary.
It was processing N-times, if there are N-Rules.

Now, it only processes one time by taking all N-rules at a time.

Conflicts:
	ui/scripts/ui/widgets/multiEdit.js