CLOUDSTACK-XXXX: Post upgrade VPN customer gateway Fields which have value as 'NONE ' are labelled now as 'label.none'
In general for description fields, it expects the value should be a string.
In widget, it parses this fields using _s().
As we use the ID of a dictionary as a value, so we need to parse it using _l().
So manually changed for the cases where it is using ID.